requirements.txt
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-spinner==0.1.10
colorama==0.4.6
coverage==7.2.3
cryptography==40.0.2
dnspython==2.3.0
Flask==2.2.3
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Markdown==3.4.3
MarkupSafe==2.1.2
numpy==1.24.3
pandas==2.0.2
pycparser==2.21
pymongo==4.3.3
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
six==1.16.0
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.4
zipp==3.15.0
